requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
SQLAlchemy==2.0.36
PyJWT==2.9.0
passlib[bcrypt]==1.7.4